Refactor: 폴더,코드스타일 수정
